refactor(gamesection): separate component
