--- service/app/src/app/_components/GameSection.tsx
@@ -1,82 +1,0 @@
-"use client"
-
-import { useSuspenseQuery } from "@tanstack/react-query"
-
-import { GameListItem } from "@/entities/game"
-import { getDefaultGame } from "@/entities/game/api/getDefaultGame"
-import { useGamePreview } from "@/entities/game/hooks/useGamePreview"
-import * as GameCard from "@/entities/game/ui/GameCard/gameCard"
-import SSRSafeSuspense from "@/shared/SSRSafeSuspense"
-interface GameSectionProps {
-  className?: string
-}
-
-import { GameSectionHeader } from "./GameSection/GameSectionHeader"
-import { GameSectionSkeleton } from "./GameSection/GameSectionSkeleton"
-
-interface GameSectionCardsProps {
-  onGameCardClick: (game: GameListItem) => void
-}
-
-const GameSectionCards = ({ onGameCardClick }: GameSectionCardsProps) => {
-  const { data: games } = useSuspenseQuery({
-    queryKey: ["defaultGames"],
-    queryFn: async () => await getDefaultGame({ cache: "force-cache" }),
-    select: (data) => data.data.games,
-    staleTime: Infinity,
-  })
-
-  return (
-    <div
-      className="flex items-center justify-between"
-      data-testid="game-section-cards"
-    >
-      {games?.map((game) => (
-        <GameCard.Root
-          key={game.gameId}
-          className="w-[178px]"
-          title={game.gameTitle}
-          onClick={() => onGameCardClick(game)}
-          aria-label="게임 카드"
-          tabIndex={0}
-        >
-          <GameCard.Image
-            src={game.gameThumbnailUrl ?? "/checker.svg"}
-            alt={game.gameTitle}
-            sizes="178px"
-            placeholder="blur"
-            blurDataURL="data:image/svg+xml;base64,PHN2ZyB3aWR0aD0iNzI3IiBoZWlnaHQ9IjQ1OSIgeG1sbnM9Imh0dHA6Ly93d3cudzMub3JnLzIwMDAvc3ZnIj48cmVjdCB3aWR0aD0iMTAwJSIgaGVpZ2h0PSIxMDAlIiBmaWxsPSIjZTVlN2ViIi8+PC9zdmc+"
-          >
-            <GameCard.Badge>{game.questionCount}문제</GameCard.Badge>
-            {game.isShared && (
-              <GameCard.Badge variant="bottom-left">공유</GameCard.Badge>
-            )}
-          </GameCard.Image>
-          <GameCard.Description>{game.gameTitle}</GameCard.Description>
-        </GameCard.Root>
-      ))}
-    </div>
-  )
-}
-
-export const GameSection = ({ className = "" }: GameSectionProps) => {
-  const { openPreview } = useGamePreview()
-
-  const handleGameCardClick = (game: GameListItem) => {
-    void openPreview(game)
-  }
-
-  return (
-    <section
-      className={`flex w-full flex-col items-center justify-center self-stretch p-0 ${className}`}
-      aria-label="게임 섹션"
-    >
-      <div className="flex min-w-[952px] flex-col gap-28">
-        <GameSectionHeader />
-        <SSRSafeSuspense fallback={<GameSectionSkeleton />}>
-          <GameSectionCards onGameCardClick={handleGameCardClick} />
-        </SSRSafeSuspense>
-      </div>
-    </section>
-  )
-}
